Helper methods moved to utils due to comments.
