+++ aea/cli/utils/package_utils.py
@@ -40,6 +40,7 @@ from aea.configurations.base import (
     DEFAULT_AEA_CONFIG_FILE,
     PackageType,
     PublicId,
+    _compute_fingerprint,
     _get_default_configuration_file_name_from_type,
 )
 from aea.configurations.loader import ConfigLoader
@@ -375,3 +376,63 @@ def validate_author_name(author: Optional[str] = None) -> str:
             )
 
     return valid_author
+
+
+def is_fingerprint_correct(package_path: Path, item_config) -> bool:
+    """
+    Validate fingerprint of item before adding.
+
+    :param package_path: path to a package folder.
+    :param item_config: item configuration.
+
+    :return: None.
+    """
+    fingerprint = _compute_fingerprint(
+        package_path, ignore_patterns=item_config.fingerprint_ignore_patterns
+    )
+    return item_config.fingerprint == fingerprint
+
+
+def register_item(ctx: Context, item_type: str, item_public_id: PublicId) -> None:
+    """
+    Register item in agent configuration.
+
+    :param ctx: click context object.
+    :param item_type: type of item.
+    :param item_public_id: PublicId of item.
+
+    :return: None.
+    """
+    logger.debug(
+        "Registering the {} into {}".format(item_type, DEFAULT_AEA_CONFIG_FILE)
+    )
+    item_type_plural = item_type + "s"
+    supported_items = getattr(ctx.agent_config, item_type_plural)
+    supported_items.add(item_public_id)
+    ctx.agent_loader.dump(
+        ctx.agent_config, open(os.path.join(ctx.cwd, DEFAULT_AEA_CONFIG_FILE), "w")
+    )
+
+
+def is_item_present(ctx: Context, item_type: str, item_public_id: PublicId) -> bool:
+    """
+    Check if item is already present in AEA.
+
+    :param ctx: context object.
+    :param item_type: type of an item.
+    :param item_public_id: PublicId of an item.
+
+    :return: boolean is item present.
+    """
+    item_type_plural = item_type + "s"
+    dest_path = Path(
+        ctx.cwd, "vendor", item_public_id.author, item_type_plural, item_public_id.name
+    )
+    # check item presence only by author/package_name pair, without version.
+    items_in_config = set(
+        map(lambda x: (x.author, x.name), getattr(ctx.agent_config, item_type_plural))
+    )
+    return (
+        item_public_id.author,
+        item_public_id.name,
+    ) in items_in_config and dest_path.exists()
